+++ packages/runtime/src/runtime/hooks/useEnsembleAction.tsx
@@ -12,6 +12,7 @@ import {
   DateFormatter,
   useApplicationContext,
   unwrapWidget,
+  useEnsembleUser,
 } from "@ensembleui/react-framework";
 import type {
   InvokeAPIAction,
@@ -90,7 +91,7 @@ export const useExecuteCode: EnsembleActionHook<
       )?.body;
     }
   }, [action, isCodeString, screen]);
-
+  const [user] = useEnsembleUser();
   const appContext = useApplicationContext();
   const onCompleteAction = useEnsembleAction(
     isCodeString ? undefined : action?.onComplete,
@@ -110,6 +111,7 @@ export const useExecuteCode: EnsembleActionHook<
             {
               ensemble: {
                 storage,
+                user,
                 formatter,
                 env: appContext?.env,
                 navigateScreen: (targetScreen: NavigateScreenAction): void =>